fix: fixed some SMALL memory issue and fixed #456
