+++ src/ui/widgets/RouteSettingsMatrix.ui
@@ -6,92 +6,131 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>927</width>
-    <height>198</height>
+    <width>600</width>
+    <height>409</height>
    </rect>
   </property>
   <property name="windowTitle">
    <string>Form</string>
   </property>
-  <layout class="QVBoxLayout" name="verticalLayout">
+  <layout class="QVBoxLayout" name="verticalLayout_2">
    <item>
-    <widget class="QLabel" name="label_3">
-     <property name="text">
-      <string>Lines start with &quot;geoip:&quot; or &quot;geosite:&quot; will have its autocompletion from geoip.dat and geosite.dat</string>
-     </property>
-    </widget>
-   </item>
-   <item>
-    <layout class="QGridLayout" name="gridLayout" rowstretch="0,1,1" columnstretch="0,1,1,1,0">
-     <item row="1" column="0">
-      <widget class="QLabel" name="label_2">
+    <layout class="QHBoxLayout" name="horizontalLayout_2" stretch="0,1">
+     <item>
+      <widget class="QLabel" name="label_4">
        <property name="text">
-        <string>IP</string>
-       </property>
-       <property name="alignment">
-        <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
+        <string>Domain Strategy</string>
        </property>
       </widget>
      </item>
-     <item row="2" column="0">
-      <widget class="QLabel" name="label">
-       <property name="text">
-        <string>Domain</string>
-       </property>
-       <property name="alignment">
-        <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
-       </property>
-      </widget>
-     </item>
-     <item row="1" column="1">
-      <layout class="QGridLayout" name="directIPLayout"/>
-     </item>
-     <item row="2" column="1">
-      <layout class="QGridLayout" name="directTxtLayout"/>
-     </item>
-     <item row="1" column="2">
-      <layout class="QGridLayout" name="proxyIPLayout"/>
-     </item>
-     <item row="2" column="2">
-      <layout class="QGridLayout" name="proxyTxtLayout"/>
-     </item>
-     <item row="1" column="3">
-      <layout class="QGridLayout" name="blockIPLayout"/>
-     </item>
-     <item row="2" column="3">
-      <layout class="QGridLayout" name="blockTxtLayout"/>
-     </item>
-     <item row="0" column="1">
-      <widget class="QLabel" name="label_80">
-       <property name="text">
-        <string>Direct</string>
-       </property>
-       <property name="alignment">
-        <set>Qt::AlignCenter</set>
-       </property>
-      </widget>
-     </item>
-     <item row="0" column="2">
-      <widget class="QLabel" name="label_81">
-       <property name="text">
-        <string>Proxy</string>
-       </property>
-       <property name="alignment">
-        <set>Qt::AlignCenter</set>
-       </property>
-      </widget>
-     </item>
-     <item row="0" column="3">
-      <widget class="QLabel" name="label_82">
-       <property name="text">
-        <string>Block</string>
-       </property>
-       <property name="alignment">
-        <set>Qt::AlignCenter</set>
-       </property>
+     <item>
+      <widget class="QComboBox" name="domainStrategyCombo">
+       <item>
+        <property name="text">
+         <string notr="true">AsIs</string>
+        </property>
+       </item>
+       <item>
+        <property name="text">
+         <string notr="true">IPIfNonMatch</string>
+        </property>
+       </item>
+       <item>
+        <property name="text">
+         <string notr="true">IPOnDemand</string>
+        </property>
+       </item>
       </widget>
      </item>
     </layout>
+   </item>
+   <item>
+    <widget class="QGroupBox" name="groupBox">
+     <property name="title">
+      <string>Route Settings</string>
+     </property>
+     <layout class="QVBoxLayout" name="verticalLayout">
+      <item>
+       <widget class="QLabel" name="label_3">
+        <property name="text">
+         <string>Lines start with &quot;geoip:&quot; or &quot;geosite:&quot; will have its autocompletion from geoip.dat and geosite.dat</string>
+        </property>
+       </widget>
+      </item>
+      <item>
+       <layout class="QGridLayout" name="gridLayout" rowstretch="0,1,1" columnstretch="0,1,1,1">
+        <item row="0" column="3">
+         <widget class="QLabel" name="label_82">
+          <property name="text">
+           <string>Block</string>
+          </property>
+          <property name="alignment">
+           <set>Qt::AlignCenter</set>
+          </property>
+         </widget>
+        </item>
+        <item row="2" column="3">
+         <layout class="QGridLayout" name="blockTxtLayout"/>
+        </item>
+        <item row="0" column="1">
+         <widget class="QLabel" name="label_80">
+          <property name="text">
+           <string>Direct</string>
+          </property>
+          <property name="alignment">
+           <set>Qt::AlignCenter</set>
+          </property>
+         </widget>
+        </item>
+        <item row="2" column="0">
+         <widget class="QLabel" name="label">
+          <property name="text">
+           <string>Domain</string>
+          </property>
+          <property name="alignment">
+           <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
+          </property>
+         </widget>
+        </item>
+        <item row="2" column="1">
+         <layout class="QGridLayout" name="directTxtLayout"/>
+        </item>
+        <item row="1" column="2">
+         <layout class="QGridLayout" name="proxyIPLayout"/>
+        </item>
+        <item row="0" column="2">
+         <widget class="QLabel" name="label_81">
+          <property name="text">
+           <string>Proxy</string>
+          </property>
+          <property name="alignment">
+           <set>Qt::AlignCenter</set>
+          </property>
+         </widget>
+        </item>
+        <item row="2" column="2">
+         <layout class="QGridLayout" name="proxyTxtLayout"/>
+        </item>
+        <item row="1" column="3">
+         <layout class="QGridLayout" name="blockIPLayout"/>
+        </item>
+        <item row="1" column="0">
+         <widget class="QLabel" name="label_2">
+          <property name="text">
+           <string>IP</string>
+          </property>
+          <property name="alignment">
+           <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
+          </property>
+         </widget>
+        </item>
+        <item row="1" column="1">
+         <layout class="QGridLayout" name="directIPLayout"/>
+        </item>
+       </layout>
+      </item>
+     </layout>
+    </widget>
    </item>
    <item>
     <layout class="QHBoxLayout" name="horizontalLayout">
